Created sendBookApi.js and cleaned App.js
